refactor: move static modules (vendor) to `Service{rovider::class`
Moving modules to service provider make nomore extra steep to create package
