+++ tests/Integrate/Commands/VendorImportCommandsTest.php
@@ -6,164 +6,77 @@ namespace System\Test\Integrate\Commands;
 
 use PHPUnit\Framework\TestCase;
 use System\Integrate\Console\VendorImportCommand;
+use System\Integrate\ServiceProvider;
 
 final class VendorImportCommandsTest extends TestCase
 {
+    private ?string $base_path;
+
+    protected function setUp(): void
+    {
+        $this->base_path = dirname(__DIR__);
+    }
+
     protected function tearDown(): void
     {
-        @unlink(__DIR__ . '/assets/copy/to/file.txt');
+        ServiceProvider::flushModule();
+        @unlink($this->base_path . '/assets/copy/to/file.txt');
     }
 
     /**
      * @test
      */
-    public function itCanExportModule(): void
+    public function itCanImport(): void
     {
-        VendorImportCommand::export([
-            '/vendor/package/database/' => '/database/',
+        $publish = new VendorImportCommand(['cli', 'vendor:import', '--tag=test'], [
+            'force' => false,
         ]);
+        $random = now()->format('YmdHis') . microtime();
 
-        VendorImportCommand::export([
-            '/vendor/package/resource/view/' => '/resourve/view/',
-        ], 'pacakge-share');
+        ServiceProvider::export(
+            path: [$this->base_path . '/assets/copy/from/file.txt' => $this->base_path . '/assets/copy/to/file.txt'],
+            tag: 'test'
+        );
+        ServiceProvider::export(
+            path: [$this->base_path . '/assets/copy/from/folder' => $this->base_path . '/assets/copy/to/folders/folder-' . $random],
+            tag: 'test'
+        );
 
-        VendorImportCommand::export([
-            '/vendor/package/resource/js/app.js' => '/resourve/js/app.js',
-        ], 'pacakge-share');
+        ob_start();
+        $exit = $publish->main();
+        ob_get_clean();
 
-        $this->assertEquals([
-            ''              => ['/vendor/package/database/' => '/database/'],
-            'pacakge-share' => [
-                '/vendor/package/resource/view/'     => '/resourve/view/',
-                '/vendor/package/resource/js/app.js' => '/resourve/js/app.js',
-            ],
-        ], VendorImportCommand::getModules());
+        $this->assertEquals(0, $exit);
+        $this->assertTrue(file_exists($this->base_path . '/assets/copy/to/file.txt'));
+        $this->assertTrue(file_exists($this->base_path . '/assets/copy/to/folders/folder-' . $random . '/file.txt'));
     }
 
     /**
      * @test
      */
-    public function itCanGetModule(): void
+    public function itCanImportWithTag(): void
     {
-        VendorImportCommand::export([
-            '/vendor/package/database/' => '/database/',
+        $publish = new VendorImportCommand(['cli', 'vendor:import', '--tag=test'], [
+            'force' => false,
+            'tag'   => 'test',
         ]);
-        VendorImportCommand::flushModule();
-
-        $this->assertEquals([], VendorImportCommand::getModules());
-    }
-
-    /**
-     * @test
-     */
-    public function itCanImportFile(): void
-    {
-        $this->assertTrue(VendorImportCommand::importFile(
-            __DIR__ . '/assets/copy/from/file.txt',
-            __DIR__ . '/assets/copy/to/file.txt'
-        ));
-
-        $this->assertTrue(file_exists(__DIR__ . '/assets/copy/to/file.txt'));
-    }
-
-    /**
-     * @test
-     */
-    public function itCanImportFileWithFolderDoestExits(): void
-    {
         $random = now()->format('YmdHis') . microtime();
-        $this->assertTrue(VendorImportCommand::importFile(
-            __DIR__ . '/assets/copy/from/file.txt',
-            __DIR__ . '/assets/copy/to/folders/folder-' . $random . '/file.txt'
-        ));
 
-        $this->assertTrue(file_exists(__DIR__ . '/assets/copy/to/folders/folder-' . $random . '/file.txt'));
-    }
+        ServiceProvider::export(
+            path: [$this->base_path . '/assets/copy/from/file.txt' => $this->base_path . '/assets/copy/to/file.txt'],
+            tag: 'test'
+        );
+        ServiceProvider::export(
+            path: [$this->base_path . '/assets/copy/from/folder' => $this->base_path . '/assets/copy/to/folders/folder-' . $random],
+            tag: 'vendor'
+        );
 
-    /**
-     * @test
-     */
-    public function itCanImportFileWithTargetExist(): void
-    {
-        file_put_contents(__DIR__ . '/assets/copy/to/file.txt', '');
+        ob_start();
+        $exit = $publish->main();
+        ob_get_clean();
 
-        $this->assertTrue(VendorImportCommand::importFile(
-            __DIR__ . '/assets/copy/from/file.txt',
-            __DIR__ . '/assets/copy/to/file.txt',
-            true
-        ));
-
-        $this->assertTrue(file_exists(__DIR__ . '/assets/copy/to/file.txt'));
-    }
-
-    /**
-     * @test
-     */
-    public function itCanNotImportFileWithTargetExist(): void
-    {
-        file_put_contents(__DIR__ . '/assets/copy/to/file.txt', '');
-
-        $this->assertFalse(VendorImportCommand::importFile(
-            __DIR__ . '/assets/copy/from/file.txt',
-            __DIR__ . '/assets/copy/to/file.txt'
-        ));
-
-        $this->assertTrue(file_exists(__DIR__ . '/assets/copy/to/file.txt'));
-    }
-
-    /**
-     * @test
-     */
-    public function itCanImportFolder(): void
-    {
-        $random = now()->format('YmdHis') . microtime();
-        $this->assertTrue(VendorImportCommand::importDir(
-            __DIR__ . '/assets/copy/from/folder',
-            __DIR__ . '/assets/copy/to/folders/folder-' . $random
-        ));
-
-        $this->assertTrue(file_exists(__DIR__ . '/assets/copy/to/folders/folder-' . $random . '/file.txt'));
-    }
-
-    /**
-     * @test
-     */
-    public function itCanImportFolderRecursing(): void
-    {
-        $random = now()->format('YmdHis') . microtime();
-        $this->assertTrue(VendorImportCommand::importDir(
-            __DIR__ . '/assets/copy/from/folder-nest',
-            __DIR__ . '/assets/copy/to/folders/folder-' . $random
-        ));
-
-        $this->assertTrue(file_exists(__DIR__ . '/assets/copy/to/folders/folder-' . $random . '/file.txt'));
-        $this->assertTrue(file_exists(__DIR__ . '/assets/copy/to/folders/folder-' . $random . '/folder/file.txt'));
-    }
-
-    /**
-     * @test
-     */
-    public function itCanImportFolderWithTargetExist(): void
-    {
-        $this->assertTrue(VendorImportCommand::importDir(
-            __DIR__ . '/assets/copy/from/folder',
-            __DIR__ . '/assets/copy/to/folder',
-            true
-        ));
-
-        $this->assertTrue(file_exists(__DIR__ . '/assets/copy/to/folder/file.txt'));
-    }
-
-    /**
-     * @test
-     */
-    public function itCaNotImportFolderWithTargetExist(): void
-    {
-        $this->assertFalse(VendorImportCommand::importDir(
-            __DIR__ . '/assets/copy/from/folder',
-            __DIR__ . '/assets/copy/to/folder'
-        ));
-
-        $this->assertTrue(file_exists(__DIR__ . '/assets/copy/to/folder/file.txt'));
+        $this->assertEquals(0, $exit);
+        $this->assertTrue(file_exists($this->base_path . '/assets/copy/to/file.txt'));
+        $this->assertFalse(file_exists($this->base_path . '/assets/copy/to/folders/folder-' . $random . '/file.txt'));
     }
 }
