+++ tests/Integrate/assets/copy/to/folders/.gitignore
